# Text format (e.g. IridasCube) parsing optimizations