--- a/src/OpenColorIO/fileformats/FileFormatSpi1D.cpp
+++ b/src/OpenColorIO/fileformats/FileFormatSpi1D.cpp
@@ -168,7 +168,7 @@ CachedFileRcPtr LocalFileFormat::read(std::istream & istream,
                 }
             }
         }
-        while (istream.good() && !StringUtils::StartsWith(headerLine,"{"));
+        while (istream.good() && !StringUtils::StartsWith(headerLine,'{'));
     }
 
     if (version == -1)