Primarily driven by a wish to increase performance of parsing .cube files. But the functions that are added or changed are used across parsing of all/most of text based formats.

With these changes, parsing "Khronos PBR Neutral" Iridas .cube file (5.4MB) on Ryzen 5950X / VS2022 Release build: **167ms -> 126ms**. Parsing the same file in CLF format: 150ms -> 137ms.

What the PR does is:
- Add locale independent `IsSpace(char)`. Somewhat similar to 3aab90d31, whitespace trimming perhaps should not be locale dependent.
- Add `IsEmptyOrWhiteSpace()` and use that inside `ParseUtils::nextline`, instead of doing `Trim(line).empty()`.
- Add `StringUtils::StartsWith(char)` and use that in various parsers that were constructing whole `std::string` object just to check for a single character.
- When building for C++17 or later, `NumberUtils` can use standard `<charconv>` `from_chars` functions (except on Apple platforms, where those are not implemented for floating point types as of Xcode 15). This has advantage of not having to deal with `errno` or locales. Saves some thread local storage accesses and function calls (e.g. on Windows `errno` is actually a function call).
  - Turns out that `NumberUtils::from_chars` was only indirectly covered by tests, and as XML parser utils tests point out,
    it should be able to handle hex floats and so on. So all that is implemented there, and explicitly covered by
    `tests/utils/NumberUtils_tests.cpp`.
  - There's a CMake setup change that adds `/Zc:__cplusplus` flag for MSVC; for backwards compat reasons it does not report proper C++ version detection defines otherwise.